chore: merge balance component and optimize its logic to use client-side fetching and refetch interval with react query.
