--- src/app/pwa/(pages)/pools/_components/balance-info.tsx
@@ -1,8 +1,0 @@
-import { getAddressBalanceAction } from '../../profile/actions'
-import Balance from './balance'
-
-export default async function PoolsBalance() {
-    const [balanceResult] = await getAddressBalanceAction()
-
-    return <Balance initialBalance={balanceResult} />
-}
